Port ScriptBuffer to the new IPC serialization format https://bugs.webkit.org/show_bug.cgi?id=268676
Reviewed by Alex Christensen.

* Source/WebCore/workers/ScriptBuffer.cpp:
(WebCore::tryConvertToShareableResourceHandle):
(WebCore::ScriptBuffer::fromIPCData):
(WebCore::ScriptBuffer::ipcData const):
* Source/WebCore/workers/ScriptBuffer.h:
* Source/WebKit/Shared/WebCoreArgumentCoders.cpp:
(IPC::tryConvertToShareableResourceHandle): Deleted.
(IPC::ArgumentCoder<WebCore::ScriptBuffer>::encode): Deleted.
(IPC::ArgumentCoder<WebCore::ScriptBuffer>::decode): Deleted.
* Source/WebKit/Shared/WebCoreArgumentCoders.h:
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:

Canonical link: https://commits.webkit.org/274051@main
