+++ Source/WebKit/Shared/WebCoreArgumentCoders.cpp
@@ -512,54 +512,4 @@ std::optional<Ref<WebCore::FragmentedSharedBuffer>> ArgumentCoder<WebCore::Fragm
     return SharedBuffer::create(static_cast<unsigned char*>(sharedMemoryBuffer->data()), bufferSize);
 }
 
-#if ENABLE(SHAREABLE_RESOURCE) && PLATFORM(COCOA)
-static std::optional<ShareableResource::Handle> tryConvertToShareableResourceHandle(const ScriptBuffer& script)
-{
-    if (!script.containsSingleFileMappedSegment())
-        return std::nullopt;
-
-    auto& segment = script.buffer()->begin()->segment;
-    auto sharedMemory = SharedMemory::wrapMap(const_cast<uint8_t*>(segment->data()), segment->size(), SharedMemory::Protection::ReadOnly);
-    if (!sharedMemory)
-        return std::nullopt;
-
-    auto shareableResource = ShareableResource::create(sharedMemory.releaseNonNull(), 0, segment->size());
-    if (!shareableResource)
-        return std::nullopt;
-
-    return shareableResource->createHandle();
-}
-#endif
-
-void ArgumentCoder<WebCore::ScriptBuffer>::encode(Encoder& encoder, const WebCore::ScriptBuffer& script)
-{
-#if ENABLE(SHAREABLE_RESOURCE) && PLATFORM(COCOA)
-    auto handle = tryConvertToShareableResourceHandle(script);
-    bool isShareableResourceHandle = !!handle;
-    encoder << WTFMove(handle);
-    if (isShareableResourceHandle)
-        return;
-#endif
-    encoder << RefPtr { script.buffer() };
-}
-
-std::optional<WebCore::ScriptBuffer> ArgumentCoder<WebCore::ScriptBuffer>::decode(Decoder& decoder)
-{
-#if ENABLE(SHAREABLE_RESOURCE) && PLATFORM(COCOA)
-    auto handle = decoder.decode<std::optional<ShareableResource::Handle>>();
-    if (UNLIKELY(!decoder.isValid()))
-        return std::nullopt;
-
-    if (*handle) {
-        if (auto buffer = WTFMove(**handle).tryWrapInSharedBuffer())
-            return WebCore::ScriptBuffer { WTFMove(buffer) };
-        return std::nullopt;
-    }
-#endif
-
-    if (auto buffer = decoder.decode<RefPtr<FragmentedSharedBuffer>>())
-        return WebCore::ScriptBuffer { WTFMove(*buffer) };
-    return std::nullopt;
-}
-
 } // namespace IPC
