+++ Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in
@@ -5635,6 +5635,18 @@ struct WebCore::ServiceWorkerImportedScript {
     String mimeType;
 };
 
+#if ENABLE(SHAREABLE_RESOURCE) && PLATFORM(COCOA)
+[CreateUsing=fromIPCData] class WebCore::ScriptBuffer {
+    std::variant<WebCore::ShareableResourceHandle, RefPtr<WebCore::FragmentedSharedBuffer>> ipcData();
+};
+#endif
+
+#if !ENABLE(SHAREABLE_RESOURCE) || !PLATFORM(COCOA)
+[CreateUsing=fromIPCData] class WebCore::ScriptBuffer {
+    RefPtr<WebCore::FragmentedSharedBuffer> ipcData();
+};
+#endif
+
 header: <wtf/RobinHoodHashTable.h>
 struct WebCore::ServiceWorkerContextData {
     std::optional<WebCore::ServiceWorkerJobDataIdentifier> jobDataIdentifier;
